Adjust the --pod-manifest-path flag to become the --config flag.
In a recent version of the script the --config flag was replaced with the
--pod-manifest-path flag that pointed to the kubelet-config.yaml file.

This can lead to a kubelet not working, so this will check for this specific
misconfiguration and adjust it back.

In newer cluster - as the statement was removed this should not happen anymore
and with version 4 we might be able to remove this workaround.